requirements.txt
streamlit>=1.35.0
gspread>=6.0.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.127.0
veryfi>=4.0.0
Pillow>=10.3.0
python-dotenv>=1.0.1
pandas>=2.2.0
